+++ Thunderstore/CHANGELOG.md
@@ -2,3 +2,4 @@
 ## Changes
 ### Bugfixes
 - Fixed a bug where dying could softlock you if you had never equipped a portal gun in a level.
+- Fixed portal guns being accessible even if they were disabled through the config menu.